Related work items: #3217937

+++ ms_contact_management/domain/model/src/main/java/co/com/bancolombia/model/newness/Newness.java
@@ -26,7 +26,7 @@ public class Newness {
     private String userCreation;
     private String responseCode;
     private String responseDescription;
-    private Integer voucher;
+    private Long voucher;
     private LocalDateTime dateFistInscription;
     private LocalDateTime dateCreation;
 }
